Set h_ifindex correctly.
We also use this index as a way to define an interface. The following
comment has been added to lldpd.h:

An interface is uniquely identified by h_ifindex, h_ifname and h_ops.
This means if an interface becomes enslaved, it will be considered as
a new interface. The same applies for renaming and we include the
index in case of renaming to an existing interface.