Use "void" instead of empty parameter list for function prototypes.
Using empty parameter list means that we don't know what are the
parameters of a function, while using "void" means that there is no
parameter for this function. Some compilers are ticky about this. See
C99 6.11.6 and 6.7.5.3 §14.

This also allows to fix a little bug: priv_ctl_create was called with
a parameter.