Get timezone info before being in the chroot.
Depending on the libc, first implicit call to tzset() can be done when
invoking syslog() the first time. In most situation, this is done when
we are in a chroot and therefore, we cannot read /etc/localtime and
get an accurate timezone information. Therefore, we invoke tzset()
before being in chroot.