Don't set MFS with MTU value.
Because there is no way to retrieve MFS value from Linux (unless we
try to set it to some arbitrary value), we did use the MTU instead.
This is incorrect, as notified in ticket #31.

Since some equipements could use this value to trigger an alert, we
prefer not to send any value. We don't seem able to detect for VLAN
support or Jumbo frames support.