Fix key/value output was incorrect when an interface name contains a dot.
When a dot was present in an interface name, the output of `lldpctl -f
keyvalue` was incorrect with something like this:

    lldp.vif2.vif4.vif5.vif6.vif7.0.via=LLDP
    lldp.vif2.vif4.vif5.vif6.vif7.0.rid=2

We fix this by using `\1` as a separator instead of `.` when building
the key. `\1` is replaced by `.` before printing.

When parsing the output, the interface name should be determined from
the first time it is used in a line like `lldp.vif2.0.via=LLDP`.