Change the way Chassis ID is assigned.
By default, chassis ID is assigned the chassis name. Once an
appropriate interface is discovered, the chassis ID is changed to the
MAC address of this interface. In most configurations, the chassis ID
is always set to the MAC address of the first interface, like this was
before this patch.