marshal: circumvent cast alignment problems using memcpy
gcc was issuing warnings like this:
 marshal.c:290:5: warning: cast increases required alignment of target type [-Wcast-align]

They are harmless because for one of them, the target has been
malloced (and therefore is safely aligned to 8 bytes) and for the
other because we know that the target location is the location of a
pointer. For those later cases, assignment is replaced by memcpy().