lldpd: fix chassis copy
When a chassis is already known, we copy the new one over the old one
and keep the old one. With the addition of management addresses as a
list, the copy was incorrect. We fix it. This is kludgy. A better way
would be to use marshaling but marshaling is not able to do inplace
copy and it is not able to restore full list (previous pointer is not
restored).