Separate daemon and client code. Provide a client library.
In `daemon/`, we have what is exclusively used by lldpd. In `client/`,
we have what is exclusively used by lldpctl and in `lib/`, we have a
library that will talk to the daemon and can be used to build other
clients. lldpctl is using this library.

This is a mega commit.

There are also some extra little changes, notably, adding a line to
each file to set "openbsd" as the default C style.