build: use a simplified autogen.sh script
The previous one was fairly complete but did not manage to generate
what was needed to get the build prepared. The bug is in autoreconf
which was calling `libtoolize` in `libevent` before doing it for the
main project. Hence, `ltmain.sh` was not found. It is still surprising
that we have such hackish build systems in 2012.