interfaces: harmonize arguments of detection functions
Some detection functions were requiring a name, other an index. We
always use a `struct netlink_interface`. The only drawback is that
those functions can only be called in the context where this structure
exists, i.e when discovering interfaces. In the case of bonding, we
used one function to detect if the interface was active. We replace
this by an unconditional replacement of the source MAC address by 0.