OpenBSD support.
With VLAN, bonding and bridge support as well. Tested with OpenBSD
5.2.

To ease porting, we also shop `netinet/if_ether.h` which is a stripped
down version of the one contained in OpenBSD (without kernel and ARP
stuff). Including correctly this header has always been a pain, even
when supporting only Linux.