lldpd: limit the maximum of neighbors per port and per protocol
lldpd memory can be exhausted by adding too many neighbors. Moreover,
the communication with lldpcli is limited to a fixed number of
neighbors (packets may not exceed 1 << 19 in size). Therefore, we put
a hard limit to the number of neighbors accepted for a given port, for
a given protocol.

The limit will be tunable in the future.

Feature suggested by Chris Yang. Closes #32.