osx: add a target to build an OSX package
We use pkgbuild and productbuild (which seem to have superseded
PackageMaker). Those tools are shipped from OSX 10.6.6. There are only
a few examples of how to use those tools with daemons. You may also
look at Jenkins or ngircd.

See `README.md` for more information on how to invoke the build of a
package.