interfaces: overwrite interface description with neighbor found
If no neighbor is found, interface name is `lldpd: no neighbor
found`. If one neighbor is found, this is `lldpd: connected to XXXXX`
and if several neighbors are found, this is `lldpd: XXX
neighbors`. Smart filter is used to count neighbors.

Currently, this is not possible to disable this. Works with FreeBSD,
OpenBSD and Linux.