linux: mangle MAC address for enslaved devices
With enslaved devices, we may end up with in a situation where we send
a packet with a source MAC address that may confuse switches. For
example:

 1. eth0, MAC X connected to switch A.
 2. eth1, MAC Y connected to switch B.
 3. Active/backup bond is created, MAC X.
 4. eth1 is active, eth0 is backup.

If we use the real MAC, eth0 will send LLDP packets with MAC X to
switch A that will learn this MAC and send packets on backup port eth0
and they will be discarded.

The solution until then was to use a zero MAC address in those
cases. However, it seems that some switches may be confused by such a
move. See #26.

Instead, we set the "local" bit to 1 if not already set. Otherwise, we
just use some arbitrary MAC address that I have on an unused 3Com
"Tornado" card. It is unlikely to be reused (100 MBps and it is in a
basement). If you happen to have an even older MAC address and are
willing to give it for lldpd, feel free to tell me. The later case
does not handle all problems but allievate the problem with switches
discarding or complaining about zero MAC address as a source.