Revert "README: remove inexact paragraph about stacking bonding/vlans/bridges"
This reverts commit ba09084d47643b630f3705242c1e6761d0412c46.

In fact, we still have limitations on how stacking is done. Bonds are
expected to be done on physical devices. Bonds on bonds are not
supported. In fact, there are some setups that cannot work
correctly. For example, let's assume the following:

 - eth0
 - eth1
 - vlan40@eth0
 - vlan41@eth1
 - vlan42@eth0
 - vlan43@eth1
 - bond0@(vlan40 + vlan41)
 - bond1@(vlan43 + vlan42)

On paper, we can still retrieve the real MAC addresses of various
stuff. However, Linux will alter the MAC addresses of the
VLAN. Moreover, depending on how bond1 is built, it can have the same
MAC address than bond0 or a different one. If MAC addresses are
propagated to the root, there may or may not have conflicts.

Putting bonds on physical interfaces should be the only sane thing to
do. While we could accomodate more complex setup, let's not try to do
something quite complex that nobody will use.

Hence the revert.