lldpcli: return success when reading commands from files
When reading commands from files, we chose to not stop on errors (the
main reason is to allow to specify nonexistent configuration files and
allow lldpd to still work after incompatible updates). However, we also
always return a failure status. Fix that.

Thanks to Brian Bidulock for spotting that. Closes #62.