lldpd: create pidfile after forking for privilege separation
Currently when privilege is enabled, the pidfile contains the PID
for the pre-fork PID. This causes any init system that relies on
pidfiles to always fail to stop the daemon. This moves creation
into the priv_init function, using an extra parameter to control
whether it gets created or not.