build: use AC_CONFIG_FILES only to generate Makefiles
Other kind of files may get partially expanded variables. For example,
`@sbindir@` is expanded to `${exec_prefix}/sbin` which won't work in a
regular configuration file.

While this could be done easily with GNU Make, using automake doesn't
allow us to use pattern rules or stuff like that. Therefore, there are a
lot of repetitions to declare the files we need to transform.