interfaces: bond listening was broken only for old kernels
With a recent kernel, we can listen to individual member. lldpd is still
listening on master to keep compatibility with older kernels and because
it is difficult to detect if the kernel gives us this ability or not. We
can't rely on version numbers as some kernels have backported
features (like 2.6.18 from RHEL 5).