lldp: when MSAP changes, send a shutdown LLPDU
Chassis ID and port ID are not supposed to change. However, it is
possible for the port ID to change (for example, if the port name
changes or if the MAC addresses change). Since they should be constant,
we "fix" that by sending a shutdown LLPDU before any broadcasting any
change.

As a provision, this also applies to chassis ID even if it is currently
not possible for it to be changed.