lldpd: consider interfaces which are not running also for RTM_DELLINK event
That way hardware->h_flags get updated, and the hardware object
is not free'd.
When the interface comes back online (with a link), the iface->flags
(and consequently hardware->h_flags) get updated.

This is a bit easier than trying to do other more complex logic,
for keeping duplicate (persistent) information.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>