lib: ability to get and query local chassis
Up until now, to get information about the local chassis, we had to get
one interface and query the associated local port, then get chassis
information out of it.

With this change, a dedicated `lldpctl_get_local_chassis()` function is
provided and returns a chassis atom. All the chassis-related properties
are moved to this kind of atom. We also add a `lldpctl_k_port_chassis`
key to get the chassis associated to a port.

Left for next commits:
 - lldpcli hasn't been updated
 - compatibility layer to query chassis-related properties from port