lldpcli: rename "unconfigure lldp custom-tlvs" to "unconfigure lldp custom-tlv"
There was a bit of confusion regarding this command.
It's documented as 'unconfigure lldp custom-tlv', when in fact it's
'unconfigure lldp custom-tlvs' (plural).

That's also my fault, since I did not update much (or take care about)
documentation.
It seems simpler to rename the command as is in the doc than change
the doc.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>