netlink: use libnl3 instead of custom netlink code
The main goal of this change is to implement a caching system for
netlink and avoiding the full scan done when a change was
triggered. Implementing a netlink cache is not as funny as it seems and
therefore, it seems just better to use libnl3.

Licensing issues are explained in README. People concerned with that
should just use dynamic linking. An embedded copy of libnl3 is also
provided, just like libevent.