interfaces: fix alias handling
And try to add some simple tests for them. The manual page says that if
an alias is provided, the default is to use the alias as description and
port name as port ID. We try to respect that.

The tests are not as complete as they could be because the MSAP changes
due to the port ID change and this makes tests more complex.