netlink: don't consider a lower interface when in another namespace
The index of an interface is specific to a namespace, don't try to
interpret anything about interfaces belonging to another namespace. This
change unbreak some scenario, like in an appropriate loop detection
because an interface from another namespace is detected (vlan100 ->
veth1 -> veth0 with same index as vlan100). However, it is not possible
to exactly detect a physical interface anymore since we don't really
know what can be on the other side of the interface (in the other
namespace). However, bridged, bonded and VLAN interfaces should be safe.