interfaces: handle correctly operation conversation of a port
When a port was regular and become an enslaved member of a bond, we
created a different port. Since we now keep the old ports around, the
old port may still attract a lot of things, like specific configuration
or VLAN. Therefore, we handle the conversion of a port from one kind to
another.

Another idea would be to not do special handling for bonds. Only old
kernels need that. We could remove that later.