build: fix build issue introduced in previous commit
$^ was used for a reason: automake would fix the paths used in
dependencies using VPATH or something similar. For some reason, this
even break with GNU make when not using OOT build. Long story short, we
use a pattern substitution to ensure that the atom files are looked up
in the correct directory.

This use of substitutions is mandated by POSIX (the 2013 version I
think). See:
 http://austingroupbugs.net/view.php?id=519

It is however believed that most make implementation had support for
this since a long time.