lib/connection: simplify lldpctl_recv() function
The final aim is to have async IO.

We're using uloop() (from OpenWrt) which is a wrapper on top
of epoll().
uloop() is a main-loop type of mechanism, which means FDs need
to be assigned to it, and then if the epoll() has some events,
it would trigger the handlers for the registered (with epoll) FDs.

With the current way liblldpctl looks like, async IO is not too
perfect.

But, regarding lldpctl_recv() it looks like it could call the
conn->recv() pointer, since it's mandatory anyway to pass it
in order to get a async-IO-ish mechanism.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>