interfaces: on Linux, clear socket error queue on error
With AF_PACKET, it is possible to get an error in the error queue even
when we didn't ask to receive them (in contrast with AF_INET where we
have to set IP_RECVERR socket option). Currently, the only error we can
receive is SOL_PACKET/PACKET_TX_TIMESTAMP, but we just want to clear any
error in the socket queue to be able to read packets (otherwise, poll()
will always return POLLERR).

Fix #215.