tests: avoid race condition when linking two namespaces
When creating a link between two namespaces, one lldpd will see the link
and send a LLDPDU before the other one. There is a slight delay for
lldpd to wait before sending the LLDPDU, but when the load is heavy, it
can be expected that the other side didn't configure the interface when
the LLDPDU is received. To solve this, just wait a bit before putting up
the second interface.