daemon: config netlink socket buf sizes at build time
I admit it would be nicer to have this configurable at run-time
or better yet, introduce a mechanism for recovering from
overrun situations via polling.

We're starting to see in certain situations that on devices
with many ports (i.e. switches with around ~50 ports) lldpd
info is not always consistent.

I'm not entirely sure that the issue is with netlink
sock buffer sizes.

But we've also recentl addressed some issues for the mstpd
daemon with netlink sock buffer sizes.
So, I'm tempted to think that it's the same issue.

We [especially I] prefer to bump sock buffer sizes where needed
and not via sysctl `net.core.rmem_default` & `net.core.wmem_default`.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>