netlink: clear IFF_SLAVE when no information on upper interface
This means that we have just been enslaved, but don't know the
master. On Linux, enslaved device are handled differently of regular
device. When we are half-enslaved (no upper interface yet), we are still
a regular interface. When we are fully enslaved, we may do nothing
because there is no local change (since flags are the same). Therefore,
clear the flag to avoid this issue.