client: built-in JSON support
Instead of relying on two different libs to serialize JSON, we provide
our own built-in serializer. This reduces the amount of build-related
code (detection of the correct library), as well as the amount of C code
with the exception of UTF-8 handling which was stolen from CCAN.

This will also make the life of some users easier.

The JSON format stays the same: quite verbose with --enable-json0
flag. And inconsistent without it.

Fix #188.
Fix #220.